Don't clobber thumbnail url on MXKAttachment
Goes with https://github.com/matrix-org/matrix-ios-kit/pull/191

This might cause a file search view to use larger thumbnails than
it did previously, although I'm not actually sure this is wired up
anywhere. If it is, I'd rather find the place they're used and
switch it to use getThumbnailUrlForSize with the size it wants.