Bug Fix - App crash: [__NSCFString replaceCharactersInRange:withString:]: nil argument
We observed some app crashes due to a nil string in a component.
According to the implementation this case should not happen because the components are removed as soon as their string is nil.
We patch here this issue by adding some logs in order to investigate it in the future.