Pushkit Support
- Bug Fix - Regression: incoming video call is ignored when the app is backgrounded.
Do not remove the call observer at each session state change. A call observer is created as soon as a session is `StoreDataReady`, it is used until all the sessions are closed. We let the MXCallManager delay the call invite handling when the concerned session is syncing (see https://github.com/matrix-org/matrix-ios-sdk/pull/366).

- Bug Fix - Regression: app crashes on logout (see wrong use of `callEventsListeners`).