Initial work on adding Dynamic Type support
Related to issue #496.

This patch is not exhaustive - it doesn't make
everything use Dynamic Type fonts.
Still, it makes most of the (frequently used) UI scalable,
which should be a nice first step and an improvement
for people with visibility problems.

Even in the generic case (no visibility problems and
font size increase set in Accessibility settings),
with this patch, the fonts appear a little larger
and more legible than before (addresses issue #1487).

This patch defines some basic fonts, listens for
a `UIContentSizeCategoryDidChangeNotification`
and maintains these fonts updated accordingly.

To let the rest of the app know about such font size changes,
we send the same `kRiotDesignValuesDidChangeThemeNotification`
which gets dispatched when people change between the Dark/Light
themes. Since the intention is similar (UI settings update & redraw),
it seemed natural to repurpose it.

Not all fonts are assigned dynamically (`widget.font = ..`),
so not all widgets use the new font values from `RiotDesignValues`.
In other places, we simply configure the Dynamic Type font inside the `.xib` file.

UI widgets which have their font defined in the `.xib` file
do not get immediately updated. Using the "Automatically Adjust Font"
property seems like an easy viable solution,
but it requires iOS>=10, so it hasn't been done.

Signed-off-by: Slavi Pantaleev <slavi@devture.com>