reskin: Improve a bit RoomVC expanded header display
RoomVC is displayed in another UINavigationController. Its bar can be translucent. Then, we just need to make sure that the main UINavigationController navigationBar is translucent when displaying the expanded header.

Note: there is a UI glitch when navigating back when the expanded header is displayed but that bug existed before (in v0.7.8)