Clarify key backup details when backup exists but is not trusted
This ports the wording changes from Riot Web
https://github.com/matrix-org/matrix-react-sdk/pull/2917.

"Use key backup" as button text feels like it's telling me I haven't set it up
at all. This changes to "Connect this device ..." which seems more reassuring
and applies for both the "add to backup" and "restore from backup" use cases.

This also updates the backup status text for the case where a backup exists, but
it's not being used by the current device so that it's much more clear that you
have a backup at least.