BF: Build: Fix build based on git tag
It was not more possible to build using tags. The reason was the fastlane git_branch method returned nil.
Internally this method executes `git symbolic-ref HEAD --short` which returned `fatal: ref HEAD is not a symbolic ref` in our case.

We now checks out the tag as a local branch.