EmojiStore: Include short name when searching for emojis
This adds the "common" short name to the list of strings to match the search text
against. Previously, only the "other" short names were included in the comparison.
This causes an issue for certain emojis like, for instance, the "Hundred Points
Symbol" where the term "100" is *only* included in the common short name. As a
result, the emoji did not previously show up when searching for "100".

Note that as a side effect, searching for "2" will now also return things such as
the "dog2" emoji. This matches the behavior in the Element Android app and also in
the emoji-mart Node.js package.

Closes: #4063

Signed-off-by: Johannes Marbach <n0-0ne+github@mailbox.org>