Fix memory leak in [RecentsDataSource dataSource:didStateChange:]
The block passed into `[MXSession listenToEventsOfTypes:onEvent:]` is stored with a strong
reference in the `listenerBlock` property of `MXSessionEventListener`. The method also
returns the listener and it is then stored into the `roomTagsListenerByUserId` dictionary
on `self`. Since a strong reference to `self` is captured in the block that means the block
retains `self` (via the capture) and `self` retains the block (via `roomTagsListenerByUserId`)
so there is a memory leak.

The bug is most easily reproduced during testing #4168 which involves closing the current
and creating a new session.

Signed-off-by: Johannes Marbach <n0-0ne+github@mailbox.org>