GH Actions: Make sure we use the latest version of MatrixKit
We use a local version of the MatrixKit podspec, MatrixKit.edited.podspec. This creates an issue in the computation of Podfile.lock. MatrixKit is not associated to a commit. `pod install` will be happy with its marketing version. It will not try to update if this marketing version matches with the one in the cache.

So delete the MatrixKit pod cache.

Generated Podfile.lock:
```
...
EXTERNAL SOURCES:
  MatrixKit:
    :podspec: MatrixKit.edited.podspec
  MatrixSDK:
    :branch: manu/test_ci_2
    :git: https://github.com/matrix-org/matrix-ios-sdk.git

CHECKOUT OPTIONS:
  MatrixSDK:
    :commit: 32614df19df3ef43e77f25a06ddb0b04e910f90b
    :git: https://github.com/matrix-org/matrix-ios-sdk.git
...
```